% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{city_geocode}
\alias{city_geocode}
\title{Geocodes of various cities.}
\format{A tibble with 334 rows and 3 variables.
\describe{
 \item{city_name}{Name of city}
 \item{lon}{Longitude}
 \item{lat}{Latitude}
}}
\source{
Amap 高德地图 2019
}
\usage{
city_geocode
}
\description{
A general dataset containing various cities' longitude and latitude information.
}
\keyword{datasets}
