% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_chunks}
\alias{col_chunks}
\title{Collect Data in Chunks}
\usage{
col_chunks(query, timevar, min, max, break_by = "weeks",
  collect = FALSE)
}
\arguments{
\item{query}{SQL query to fetch from database.}

\item{timevar}{Time-dependent variable to be splitted.}

\item{min}{Mininum or start date.}

\item{max}{Maximum or end date.}

\item{break_by}{A character string, containing one of "day", "week", "month", "quarter" or "year".}

\item{collect}{If true, retrieves data into a local tibble.}
}
\value{
Data in chunks in list form.
}
\description{
Collect Data in Chunks
}
\examples{
\dontrun{
src <- tbl(conn, "some_table")
src \%>\% col_chunks(register_time, "2019-01-01", "2019-12-31")
}
}
