% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{county_level}
\alias{county_level}
\title{Living Quality Index of Various Counties.}
\format{A tibble with 3240 rows and 8 variables:
\describe{
 \item{province_id}{省份 ID}
 \item{province_name}{省份}
 \item{city_id}{城市 ID}
 \item{city_name}{城市}
 \item{county_id}{县级 ID}
 \item{county_name}{县级}
 \item{county_level}{县级等级，A 为最好，E 为最差}
 \item{update_date}{最后更新日期}
}}
\source{
Anjuke 安居客 2017
}
\usage{
county_level
}
\description{
An in-house dataset containing various counties' information.
}
\keyword{datasets}
