% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_sheet}
\alias{download_sheet}
\title{Fetch Sheet from Source}
\usage{
download_sheet(sheet, source = "sheetlabs")
}
\arguments{
\item{sheet}{API to fetch from Sheetlabs}

\item{source}{Key to fetch from \code{Consul} to obtain credentials. Default is 'Sheetlabs'.}
}
\value{
A data frame if successful. A list of error message and code if failure
}
\description{
Fetch sheet from online API. Currently only supports Sheetlabs.
For more info visit https://sheetlabs.com/
}
