% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est_some_conn}
\alias{est_some_conn}
\alias{est_mysql_conn}
\alias{est_pgres_conn}
\alias{est_mongo_conn}
\title{Establish Connection to Common Databases}
\usage{
est_some_conn(db, drv, params)

est_mysql_conn(db, drv = RMariaDB::MariaDB())

est_pgres_conn(db, drv = RPostgreSQL::PostgreSQL())

est_mongo_conn(db)
}
\arguments{
\item{db}{Database to connect to}

\item{drv}{Database driver}

\item{params}{See \code{?get_params}}
}
\value{
Database connection if successful.
}
\description{
Fetch credentials from Consul K/V store and establish connection to database.

In order for this function to work, one will first need to register parameters in Consul.
See \code{?register_params}
}
\examples{
\dontrun{
est_mysql_conn('mysql_database')
est_pgres_conn('postgres_database')
est_mongo_conn('mongo_database')
}
}
