% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_params}
\alias{get_params}
\title{Required Consul Parameters (Keys) for Various Functions}
\usage{
get_params(func)
}
\arguments{
\item{func}{Function to make connection (e.g \code{est_mysql_conn})}
}
\value{
Vector. Parameters for function requested.
NULL if function does not require to retrieve key from Consul.
}
\description{
Parameters or keys needed for functions to establish connections.
}
\examples{
\dontrun{

# see some examples
ls(name = "package:knowboxr", pattern = "est")
get_params(est_mongo_conn)
}
}
