% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{register_params}
\alias{register_params}
\title{Register Parameters on Consul}
\usage{
register_params(id, key)
}
\arguments{
\item{id}{Consul identifier}

\item{key}{Key to register}
}
\value{
Logical. TRUE indicates success. FALSE indicated failure.
}
\description{
Register keys with dummy values into Consul KV store.

See https://www.consul.io/ for more about Consul.
}
\examples{
\dontrun{
register_params("some_api", "endpoint")

# See ?get_params
register_params("some_mysql_database", get_params(est_mysql_conn))
}
}
