% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reverse_proxy}
\alias{reverse_proxy}
\title{Do Reverse Proxy}
\usage{
reverse_proxy(conn)
}
\arguments{
\item{conn}{Proxy name (key) to fetch from \code{Consul}.}
}
\value{
An R6 object generated from \code{Processx}
}
\description{
Create a reverse proxy as a background process. Concept explained in details see here
https://unix.stackexchange.com/questions/46235/how-does-reverse-ssh-tunneling-work.
}
