% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_params}
\alias{get_params}
\title{Required Parameters for Various Functions}
\usage{
get_params(type)
}
\arguments{
\item{type}{Which kind of connection? Pick one of the following:
\itemize{
\item reverse_proxy
\item \code{mysql}
\item \code{pgres}
\item \code{mongo}
}}
}
\value{
Parameters for type requested. NULL if type not found.
}
\description{
Parameters needed for functions to establish connections.
}
