% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{register_params}
\alias{register_params}
\title{Register parameters on Consul}
\usage{
register_params(folder, key)
}
\arguments{
\item{folder}{Prefix name to register}

\item{key}{Key to register}
}
\value{
Logical. TRUE indicates success. FALSE indicated failure.
}
\description{
Register keys with dummy values into Consul KV store.

See https://www.consul.io/ for more about Consul.
}
\examples{
\dontrun{
register_params("some_api", "endpoint")
register_params("some_mysql_database", get_params("mysql"))
}
}
